stackblog/requirements.txt
Django==1.8
django-crispy-forms==1.5.2
django-extensions==1.5.6
django-faker==0.2
django-markupfield==1.3.5
fake-factory==0.5.2
Markdown==2.6.2
Pillow==2.9.0
six==1.9.0
wheel==0.24.0
